update and add tcpstate
